Fine tuning
Disable if ctrl-key pressed
Increase range in which snap occurs
Prevent image buttons appearing and BG color changing
Add fractional snap points
Move API method to more appropriate class
